# backend/openapi.json
{
  "components": {
    "schemas": {
      "Actualization-Input": {
        "description": "Фактические значения строк Кэш-фло по месяцам.\n\n``actual_until`` — индекс последнего актуализированного месяца (``-1`` — актуализация\nотсутствует). ``actuals`` — фактические значения листовых строк Кэш-фло\n(код → ряд по месяцам); применяются к периодам ``t <= actual_until``.",
        "properties": {
          "actual_until": {
            "default": -1,
            "title": "Actual Until",
            "type": "integer"
          },
          "actuals": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                    "type": "string"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Actuals",
            "type": "object"
          }
        },
        "title": "Actualization",
        "type": "object"
      },
      "Actualization-Output": {
        "description": "Фактические значения строк Кэш-фло по месяцам.\n\n``actual_until`` — индекс последнего актуализированного месяца (``-1`` — актуализация\nотсутствует). ``actuals`` — фактические значения листовых строк Кэш-фло\n(код → ряд по месяцам); применяются к периодам ``t <= actual_until``.",
        "properties": {
          "actual_until": {
            "default": -1,
            "title": "Actual Until",
            "type": "integer"
          },
          "actuals": {
            "additionalProperties": {
              "items": {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              "type": "array"
            },
            "title": "Actuals",
            "type": "object"
          }
        },
        "title": "Actualization",
        "type": "object"
      },
...
      "BreakEvenOut": {
        "properties": {
          "break_even_revenue": {
            "items": {
              "anyOf": [
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Break Even Revenue",
            "type": "array"
          },
          "margin_of_safety": {
            "items": {
              "anyOf": [
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Margin Of Safety",
            "type": "array"
          }
        },
        "required": [
          "break_even_revenue",
          "margin_of_safety"
        ],
        "title": "BreakEvenOut",
        "type": "object"
      },
      "CalcResponse": {
        "properties": {
          "actualized_cashflow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatementOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "balance": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "break_even": {
            "$ref": "#/components/schemas/BreakEvenOut"
          },
          "cashflow": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "cashflow_variance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatementOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "income": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "metrics": {
            "$ref": "#/components/schemas/MetricsOut"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "profit_use": {
            "$ref": "#/components/schemas/StatementOut"
          },
          "ratios": {
            "$ref": "#/components/schemas/RatiosOut"
          },
          "valuation": {
            "$ref": "#/components/schemas/ValuationOut"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "engine_version",
          "n",
          "income",
          "cashflow",
          "balance",
          "profit_use",
          "metrics",
          "ratios",
          "break_even",
          "valuation",
          "warnings"
        ],
        "title": "CalcResponse",
        "type": "object"
      },
...
      "Company-Input": {
        "properties": {
          "starting_balance": {
            "$ref": "#/components/schemas/StartingBalance-Input",
            "default": {
              "cash": "0",
              "debt": "0",
              "finished_goods": "0",
              "fixed_assets_net": "0",
              "foreign_monetary": "0",
              "paid_in_capital": "0",
              "payables": "0",
              "raw_materials": "0",
              "receivables": "0",
              "retained_earnings": "0"
            }
          }
        },
        "title": "Company",
        "type": "object"
      },
...
      "Environment-Input": {
        "properties": {
          "currencies": {
            "items": {
              "$ref": "#/components/schemas/Currency"
            },
            "title": "Currencies",
            "type": "array"
          },
          "fx_open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "1",
            "title": "Fx Open"
          },
          "fx_rate": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ]
            },
            "title": "Fx Rate",
            "type": "array"
          },
          "inflation": {
            "items": {
              "$ref": "#/components/schemas/InflationGroup-Input"
            },
            "title": "Inflation",
            "type": "array"
          },
          "taxes": {
            "items": {
              "$ref": "#/components/schemas/Tax-Input"
            },
            "title": "Taxes",
            "type": "array"
          }
        },
        "title": "Environment",
        "type": "object"
      },
...
      "Financing-Input": {
        "properties": {
          "auto_financing": {
            "$ref": "#/components/schemas/AutoFinancing-Input",
            "default": {
              "annual_rate": "0.18",
              "enabled": false,
              "min_balance": "0"
            }
          },
          "common_shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "0",
            "title": "Common Shares"
          },
          "deposits": {
            "items": {
              "$ref": "#/components/schemas/Deposit-Input"
            },
            "title": "Deposits",
            "type": "array"
          },
          "dividends": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                }
              ]
            },
            "title": "Dividends",
            "type": "array"
          },
          "equity": {
            "items": {
              "$ref": "#/components/schemas/EquityInjection-Input"
            },
            "title": "Equity",
            "type": "array"
          },
          "leases": {
            "items": {
              "$ref": "#/components/schemas/Lease-Input"
            },
            "title": "Leases",
            "type": "array"
          },
          "loans": {
            "items": {
              "$ref": "#/components/schemas/Loan-Input"
            },
            "title": "Loans",
            "type": "array"
          }
        },
        "title": "Financing",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestmentPlan-Input": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/Asset-Input"
            },
            "title": "Assets",
            "type": "array"
          }
        },
        "title": "InvestmentPlan",
        "type": "object"
      },
      "InvestmentPlan-Output": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/Asset-Output"
            },
            "title": "Assets",
            "type": "array"
          }
        },
        "title": "InvestmentPlan",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "MetricsOut": {
        "properties": {
          "dpb_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dpb Months"
          },
          "irr_annual": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irr Annual"
          },
          "npv": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Npv",
            "type": "string"
          },
          "pb_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb Months"
          },
          "peak_financing_need": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Financing Need"
          },
          "pi": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pi"
          },
          "pv_investments": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pv Investments"
          }
        },
        "required": [
          "npv"
        ],
        "title": "MetricsOut",
        "type": "object"
      },
...
      "OperatingPlan-Input": {
        "properties": {
          "direct_costs": {
            "items": {
              "$ref": "#/components/schemas/DirectCostLine-Input"
            },
            "title": "Direct Costs",
            "type": "array"
          },
          "fixed_costs": {
            "items": {
              "$ref": "#/components/schemas/FixedCostLine-Input"
            },
            "title": "Fixed Costs",
            "type": "array"
          },
          "production": {
            "items": {
              "$ref": "#/components/schemas/ProductionLine-Input"
            },
            "title": "Production",
            "type": "array"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "title": "Products",
            "type": "array"
          },
          "sales": {
            "items": {
              "$ref": "#/components/schemas/SalesLine-Input"
            },
            "title": "Sales",
            "type": "array"
          }
        },
        "title": "OperatingPlan",
        "type": "object"
      },
...
      "ProjectHeader": {
        "description": "Заголовок/паспорт проекта.",
        "properties": {
          "duration_months": {
            "default": 12,
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Duration Months",
            "type": "integer"
          },
          "name": {
            "default": "Без названия",
            "title": "Name",
            "type": "string"
          },
          "start_date": {
            "default": "2026-01-01",
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "title": "ProjectHeader",
        "type": "object"
      },
      "ProjectModel-Input": {
        "description": "Полная модель проекта.",
        "properties": {
          "actualization": {
            "$ref": "#/components/schemas/Actualization-Input",
            "default": {
              "actual_until": -1,
              "actuals": {}
            }
          },
          "company": {
            "$ref": "#/components/schemas/Company-Input",
            "default": {
              "starting_balance": {
                "cash": "0",
                "debt": "0",
                "finished_goods": "0",
                "fixed_assets_net": "0",
                "foreign_monetary": "0",
                "paid_in_capital": "0",
                "payables": "0",
                "raw_materials": "0",
                "receivables": "0",
                "retained_earnings": "0"
              }
            }
          },
          "environment": {
            "$ref": "#/components/schemas/Environment-Input",
            "default": {
              "currencies": [
                {
                  "code": "RUB",
                  "name": "Российский рубль"
                }
              ],
              "fx_open": "1",
              "fx_rate": [],
              "inflation": [],
              "taxes": []
            }
          },
          "financing": {
            "$ref": "#/components/schemas/Financing-Input",
            "default": {
              "auto_financing": {
                "annual_rate": "0.18",
                "enabled": false,
                "min_balance": "0"
              },
              "common_shares": "0",
              "deposits": [],
              "dividends": [],
              "equity": [],
              "leases": [],
              "loans": []
            }
          },
          "header": {
            "$ref": "#/components/schemas/ProjectHeader",
            "default": {
              "duration_months": 12,
              "name": "Без названия",
              "start_date": "2026-01-01"
            }
          },
          "investment_plan": {
            "$ref": "#/components/schemas/InvestmentPlan-Input",
            "default": {
              "assets": []
            }
          },
          "operating_plan": {
            "$ref": "#/components/schemas/OperatingPlan-Input",
            "default": {
              "direct_costs": [],
              "fixed_costs": [],
              "production": [],
              "products": [],
              "sales": []
            }
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettings-Input",
            "default": {
              "discount_rate_annual": "0.15",
              "inflation_direct": "0",
              "inflation_general": "0",
              "inflation_sales": "0",
              "inflation_wages": "0",
              "inventory_method": "average",
              "liquidation_recovery_rate": "0",
              "min_cash_balance": "0",
              "payroll_contribution_rate": "0",
              "production_cycle_months": 0,
              "profit_tax_benefit_share": "0",
              "profit_tax_rate": "0.20",
              "property_tax_rate": "0",
              "sales_tax_rate": "0",
              "terminal_growth_rate": "0",
              "valuation_earnings_multiple": "0",
              "vat_basis": "shipment",
              "vat_rate": "0"
            }
          }
        },
        "title": "ProjectModel",
        "type": "object"
      },
      "ProjectModel-Output": {
        "description": "Полная модель проекта.",
        "properties": {
          "actualization": {
            "$ref": "#/components/schemas/Actualization-Output",
            "default": {
              "actual_until": -1,
              "actuals": {}
            }
          },
          "company": {
            "$ref": "#/components/schemas/Company-Output",
            "default": {
              "starting_balance": {
                "cash": "0",
                "debt": "0",
                "finished_goods": "0",
                "fixed_assets_net": "0",
                "foreign_monetary": "0",
                "paid_in_capital": "0",
                "payables": "0",
                "raw_materials": "0",
                "receivables": "0",
                "retained_earnings": "0"
              }
            }
          },
          "environment": {
            "$ref": "#/components/schemas/Environment-Output",
            "default": {
              "currencies": [
                {
                  "code": "RUB",
                  "name": "Российский рубль"
                }
              ],
              "fx_open": "1",
              "fx_rate": [],
              "inflation": [],
              "taxes": []
            }
          },
          "financing": {
            "$ref": "#/components/schemas/Financing-Output",
            "default": {
              "auto_financing": {
                "annual_rate": "0.18",
                "enabled": false,
                "min_balance": "0"
              },
              "common_shares": "0",
              "deposits": [],
              "dividends": [],
              "equity": [],
              "leases": [],
              "loans": []
            }
          },
          "header": {
            "$ref": "#/components/schemas/ProjectHeader",
            "default": {
              "duration_months": 12,
              "name": "Без названия",
              "start_date": "2026-01-01"
            }
          },
          "investment_plan": {
            "$ref": "#/components/schemas/InvestmentPlan-Output",
            "default": {
              "assets": []
            }
          },
          "operating_plan": {
            "$ref": "#/components/schemas/OperatingPlan-Output",
            "default": {
              "direct_costs": [],
              "fixed_costs": [],
              "production": [],
              "products": [],
              "sales": []
            }
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettings-Output",
            "default": {
              "discount_rate_annual": "0.15",
              "inflation_direct": "0",
              "inflation_general": "0",
              "inflation_sales": "0",
              "inflation_wages": "0",
              "inventory_method": "average",
              "liquidation_recovery_rate": "0",
              "min_cash_balance": "0",
              "payroll_contribution_rate": "0",
              "production_cycle_months": 0,
              "profit_tax_benefit_share": "0",
              "profit_tax_rate": "0.20",
              "property_tax_rate": "0",
              "sales_tax_rate": "0",
              "terminal_growth_rate": "0",
              "valuation_earnings_multiple": "0",
              "vat_basis": "shipment",
              "vat_rate": "0"
            }
          }
        },
        "title": "ProjectModel",
        "type": "object"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "default": "",
            "title": "Full Name",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "organization_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "StatementOut": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineOut"
            },
            "title": "Lines",
            "type": "array"
          }
        },
        "required": [
          "lines"
        ],
        "title": "StatementOut",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Расчёт финансовой модели предприятия (отчёты, показатели, коэффициенты).",
    "title": "Финансовая модель — API",
    "version": "0.9.9"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Вход по email и паролю → токен доступа.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Данные текущего пользователя.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Регистрация: создаёт пользователя, его организацию и членство (owner).",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/calculate": {
      "post": {
        "description": "Рассчитать проект: вернуть отчёты, показатели эффективности и коэффициенты.",
        "operationId": "calculate_api_v1_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectModel-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalcResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate",
        "tags": [
          "calc"
        ]
      }
    },
...
    "/api/v1/templates/{template_id}": {
      "get": {
        "description": "Готовая модель шаблона (для создания проекта на её основе).",
        "operationId": "template_api_v1_templates__template_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModel-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Template",
        "tags": [
          "calc"
        ]
      }
    },
...
  }
}